Topic 1 - Excercise 1
Create a file called index.html with the correct doctype and some random content.

Add a stylesheet to the HTML file and use it to center the texts of all section elements.

Add a hidden section with Hello World inside of it.

When the page finished loading the section must fade in