Excerise 1
Create a Movie Class

Instantiate some of your favorite movies and play with them in the console.

Create an Actor class

Create a class called EventEmitter:
The on method will receive a eventName and a callback or listener that will be executed each time a that event is triggered.
The emit method will trigger events to be consumed by other functions or objects.
The off method will delete previously defined event listeners.